django/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv requirements > requirements.txt
#

-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
attrs==22.2.0 ; python_version >= '3.6'
autopep8==2.0.0
black==22.12.0
boto3==1.26.47
botocore==1.29.47 ; python_version >= '3.7'
cachecontrol==0.12.11 ; python_version >= '3.6'
cachetools==5.2.1 ; python_version ~= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
cffi==1.15.1
cfgv==3.3.1 ; python_full_version >= '3.6.1'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
distlib==0.3.6
dj-rest-auth==2.2.6
django==3.2.12
django-cors-headers==3.13.0
django-crontab==0.7.1
django-debug-toolbar==3.8.1
django-filter==22.1
django-silk==5.0.2
django-storages==1.13.2
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
emoji==2.2.0
exceptiongroup==1.1.0 ; python_version < '3.11'
filelock==3.9.0 ; python_version >= '3.7'
firebase-admin==6.0.1
flake8==5.0.4
google-api-core[grpc]==2.11.0 ; platform_python_implementation != 'PyPy'
google-api-python-client==2.72.0
google-auth==2.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2 ; python_version >= '3.7'
google-cloud-firestore==2.8.0 ; platform_python_implementation != 'PyPy'
google-cloud-storage==2.7.0 ; python_version >= '3.7'
google-crc32c==1.5.0 ; python_version >= '3.7'
google-resumable-media==2.4.0 ; python_version >= '3.7'
googleapis-common-protos==1.58.0 ; python_version >= '3.7'
gprof2dot==2022.7.29 ; python_version >= '2.7'
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.0.4
httplib2==0.21.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
identify==2.5.12 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
importlib-metadata==4.2.0 ; python_version < '3.8'
inflection==0.5.1 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
itypes==1.2.0
jinja2==3.1.2 ; python_version >= '3.7'
jmespath==1.0.1 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
mccabe==0.7.0 ; python_version >= '3.6'
msgpack==1.0.4
mypy-extensions==0.4.3
nodeenv==1.7.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
packaging==23.0 ; python_version >= '3.7'
pathspec==0.10.3 ; python_version >= '3.7'
pillow==9.4.0
platformdirs==2.6.2 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
pre-commit==2.21.0
proto-plus==1.22.2 ; python_version >= '3.6'
protobuf==4.21.12 ; python_version >= '3.7'
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1 ; python_version >= '3.6'
pycparser==2.21
pyflakes==2.5.0 ; python_version >= '3.6'
pyjwt==2.6.0 ; python_version >= '3.7'
pyparsing==3.0.9 ; python_version >= '3.1'
pytest==7.2.0 ; python_version >= '3.7'
pytest-django==4.5.2
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.7
pyyaml==6.0 ; python_version >= '3.6'
requests==2.28.1 ; python_version >= '3.7' and python_version < '4'
rsa==4.9 ; python_version >= '3.6'
ruamel.yaml==0.17.21 ; python_version >= '3'
ruamel.yaml.clib==0.2.7 ; python_version < '3.11' and platform_python_implementation == 'CPython'
s3transfer==0.6.0 ; python_version >= '3.7'
setuptools==65.6.3 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3 ; python_version >= '3.5'
tomli==2.0.1 ; python_full_version < '3.11.0a7'
typed-ast==1.5.4 ; python_version < '3.8' and implementation_name == 'cpython'
typing-extensions==4.4.0 ; python_version < '3.10'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
virtualenv==20.16.2 ; python_version >= '3.6'
zipp==3.11.0 ; python_version >= '3.7'
